requirements.txt
python-dotenv
azure-identity
azure-ai-translation-document
azure-core
openai
bs4
beautifulsoup4
lxml
msgraph-sdk

azure-search-documents==11.6.0b4
azure-storage-blob
opencensus-ext-azure

streamlit
streamlit_extras
semantic_kernel

sqlalchemy
pandas

install msgraph-core
install msgraph-sdk